Add submodule support in GitDeltaManifest
Summary: The current implementation of GitDeltaManifest treats submodule entry in a Git tree as a regular blob. This is problematic because if we try to create a Git bundle out of this entry then it will fail because there is no such blob with that particular ObjectId. I demonstrated this scenario through a new integration test without the code fix: P870454143. After the code change, this integration test works as expected.

Reviewed By: mitrandir77

Differential Revision: D50885402

fbshipit-source-id: 2366ad4ed2af91dd8551b750f91b45c15a06553d