Optimize delta computation by adjusting chunk length based on file size
Summary:
The current implementation of computing delta for a base and target file suffers quite a bit as the size of the file increases. Even for files that are around 1MB, it takes more than 10 minutes just to compute delta. The reason is that regardless of the size of the file, the granularity of the delta is always an individual byte. For large files, this becomes unnecessarily expensive without yielding any meaningful benefit. This diff adds capability in the delta mechanism to automatically adjust the granularity of the delta (i.e. the chunk size) based on the size of the file.

With these changes, we can now compute the delta very quickly even for 50MB+ files. Added unit tests to validate the same.

Reviewed By: mitrandir77

Differential Revision: D50880055

fbshipit-source-id: 242b4b7073b0ea4295c8bfd6021b5afda2720c59