sharded_map_v2: introduce ShardedMapV2
Summary:
ShardedMapV2 is the same as ShardedMap except that it doesn't compress small subtrees into terminal nodes, instead it relies purely on inlining to solve the problem of having too many small blobs.

Each ShardedMapV2Node has a conceptual weight which is defined as the sum of weights of all its inlined children, plus the count of its non-inlined children, plus one if it contains a value itself.

To figure out which of a node's children are going to be inlined and which will not:
1) Recursively figure out inlining for each child's subtree.
2) Assume that all children are going to not be inlined and calculate the weight.
3) Iterate over children in order and for each check if inlining them will not make
the weight of the node go beyond SHARDED_MAP_V2_WEIGHT_LIMIT. If so inline them,
otherwise store them in a separate blob and store their id.

This guarantees that the size of individual blobs will not grow too large, and should avoid creating too many small blobs in most cases. In particular, subtrees that would've have become a terminal node in ShardedMap will all be inlined in ShardedMapV2, with the added upside that they could potentially be stored inlined in their parent.

Reviewed By: gustavoavena

Differential Revision: D50795936

fbshipit-source-id: cd19760743cc83fddb36be96d60371e0737308ab