Prefetch the object stream before writing to the packfile/bundle
Summary:
To avoid holding too much in memory, the input for the `PackWriter` and `BundleWriter` is a stream of `PackItemObjects`. We evaluate this stream sequentially since the write needs to happen in-order one after the other. However, this also means that we delay fetching the objects from storage until we actually have to write them leading to sequential IO. This diff adds support for configurable prefetching where the write still happens sequentially but we prefetch the objects ahead of time so that the IO can happen concurrently.

I used this optimization with concurrency=10,000 while cloning aosp/manifest repo in production and its runtime reduced by 50% (from 20-22 minutes to ~10 minutes)

Reviewed By: mitrandir77

Differential Revision: D50880889

fbshipit-source-id: 49714c72b6d9ed1d2eebe519cf2035493280fcc6