enable printing of the Client Correlator in the logs for
Summary:
enable printing of the Client Correlator in the logs for background commands

It has been really really hard to debug Commit Cloud issues for years. Now we will change that.

The first step is enabling printing of the Client Correlator for background
commands. This way, looking into "hg rage" you will be able to easily match a
command with the telemetry in the Dev Command Timers scuba table!

Moreover, we are going to instrument the Commitcloud Updates and Commit Cloud
Requests scuba tables with Client Request Info as well in the following
diffs, so then it will be just awesome.

Reviewed By: zzl0

Differential Revision: D50967834

fbshipit-source-id: b5f864cb1971630d57b007fa698216fdc8eca231