Make top bar sticky
Summary:
The top bar doesn't take up all that much room, but it's fairly valuable options.

Let's make it position: sticky so it stays at the top at all times.

Sticky allows it to take up as much room as necessary in the non-scrolled state, which is what we want so that if the bar wraps or takes variable height, it still pushes down the rest of the content correctly.

Reviewed By: zats

Differential Revision: D50993794

fbshipit-source-id: 78cffc42cddf29b46c463e65bb55489804d988fc