Make update box resize to multiple lines
Summary:
The update message box for adding a message when resubmitting was only one line long. Let's make it multiple lines long so you can more easily write longer messages.

It still defaults to one line so that it's not really in your way if you don't add an update message. But it then expands to fit all the content as you type.

Reviewed By: muirdm

Differential Revision: D50820374

fbshipit-source-id: e59ce6f5eade052075a70ca9bb4b54cf53f6780d