add a test showing we can ratelim 20pct of subset of identities
Summary:
We didn't have something showing that before.

Our ratelimitng right now is configured this way that if it's set for a big tier, and we go over the limit, all hosts from that tier will be subject to have their requests dropped.

We *can*, but we don't configure it, to consistently drop just a subset of requests affecting only part of a tier, and leaving rest of it completely unaffected.

Reviewed By: mitrandir77

Differential Revision: D51034139

fbshipit-source-id: 60c37886333902b2a9b312ec861e2115ce2b3136