repo_import: require large-only mapping when importing into a large repo
Summary:
When importing into a large repo, we want to avoid the merge commit having to backsync all of the imported repo's commits when it backsyncs the merge.  This is only possible if we mark all of the merged in commits as "not-sync".  To do this, we need the name of a mapping that does not include any of the small repos, so that we can write that into the large repo's mapping table.

Note that it's not possible to perform a repo import that crosses megarepo boundaries.  To import into a location mapped to a small repo, you must import into the small repo.

Reviewed By: mitrandir77

Differential Revision: D51023858

fbshipit-source-id: 9ba58bc8f470f27714cbcb7c0aca9a5591005f8c