commit_rewriting: add methods for directly adding a large commit version with no mappings
Summary:
Currently, if we want to mark a commit as not-synced in the large repo, we need to have a small repo available to do this.  However, this is technically unnecessary, as we are only really operating on the large repo.

Introduce a way to create a large repo commit version without a small repo.

Reviewed By: mitrandir77

Differential Revision: D51023859

fbshipit-source-id: 431d404c183b92c752c61a5c658dd42f0c929540