repo_import: remove warning about multi-head repos
Summary:
This warning was added in D35608687 to warn of the dangers of importing from multi-headed repos.  The primary issue was that the imported used whatever the last commit that was encountered in the topo-sort as the merge source, which meant an arbitrary branch or tag could be merged.

This was fixed, however in D35600879 and tested with a multi-head repo in D35606582, so the warning is unnecessary and misleading.  Remove it.

Reviewed By: mitrandir77

Differential Revision: D51023855

fbshipit-source-id: 77bc12d097605a35de0cca4301ab0a58ce614d95