symlinks: fix symlinks containing paths composed of other symlinks on Windows
Summary:
There was an issue on EdenFS when trying to determine the type of a symlink: if the target path contained symlinks in its path EdenFS incorrectly would try to internally read the unresolved path, which it was incapable of.

For instance, if the directory structure looked somewhat like this:
```
├── a -> b/y
├── b -> x
└── x
    └── y
        └── z
```
trying to show the contents of `a` would fail as EdenFS would mark that symlink as a file instead of a directory symlink.

That is fixed by this diff by adding a method for resolving the target before internally trying to read the target of a symlink.

Reviewed By: xavierd

Differential Revision: D50793884

fbshipit-source-id: 920a0b4ee6f4f1b9a9fd1143608f2d42b5ca2299