Add current stack base as a suggested rebase destination
Summary: In the list of suggested rebase locations, let's add the current stack base. This is the public commit at the bottom of your current stack. We can look this up easily in the commit tree, which allows us to also add the label onto an existing other set of labels. This way, if your current stack base is also remote/main, you see "Current Stack Base, remote/main" as the suggestion.

Reviewed By: muirdm

Differential Revision: D50855683

fbshipit-source-id: 56bba5345bb8fc6c8115a8d0fe4d9780a7f1c047