Fetch additional files in commit info view
Summary:
We don't fetch ALL files for each commit in our bulk commit fetch, rather just the first 25. But if you look at the commit in the commit info sidebar, you probably want to be able to see all the files and scroll through additional pages.

So let's fetch just the single viewed commit's additional files when you look at it. This allows the additional pages to be navigated.

Reviewed By: muirdm

Differential Revision: D50982086

fbshipit-source-id: b8461c0514d1a8dc71ea059f240696e4d7d5a191