Return early with bail!
Summary:
I was working on the codebase and wanted to change this if-else to an early return, but I didn't want to to do that on the main diff where I change the business logic, so I split it here.

Early returns make the code more readable and easier to reason with, as you can make assumptions after the if blocks.

Reviewed By: clara-9

Differential Revision: D51107887

fbshipit-source-id: 3129b201cfd80592d048783efcea41eac83a43c4