Allow reverting added files in the head commit
Summary:
Added files are a bit weird with file actions. It usually doesn't make sense to revert them, except in the special case that it's earlier in the same stack. So, we can add in a revert button for added files when they happen to viewed for a head comparison.

This doesn't apply for arbitrary committed files, since your local copy won't have the file, so it would try to revert *before* the file was added, which is a no-op. Only for the head commit does it even really make sense to revert.

For uncommitted changes, we don't deal with reverting, and instead deal with tracking/untracking.

Reviewed By: muirdm

Differential Revision: D50911946

fbshipit-source-id: d6bb32284ba69ba3ded5e8528a733d606e63669f