Support viewing diff views of deleted files
Summary:
Opening a diff view in vscode for a deleted file didn't work, because it would try to use a URI for a file that doesn't exist, which vscode fails on.

We can work around this by having a new content provider for deleted files. If we detect the right side of a comparison is against head, and the file doens't exist, then we just use this special uri scheme for the deleted file. This provider just says all such files are `''`, so the comparison acts like it deleted everything.

Reviewed By: muirdm

Differential Revision: D50911144

fbshipit-source-id: 50a8cc7b77e4fd0db7b1861c0334f7e4c115dee6