Add IconStack
Summary:
Add a component that stacks two icons on top of each other. The second icon is half size and nestled into the bottom right corner.

There's a few codicons that already do this. This component lets us play more easily with various other new combinations.

For example, we can easily now put a blue dot in the corner of an icon to upsell its use / indicate that it's applicable now.

For starters, this diff changes the icon that indicates that you have unsaved message changes. It used to just be a large circle, which doesn't expain anything.

This icon stacking also uses a clip-path to cut out some space for the stacked icon. This way, the stacked icon has a little gap between it and the bottom icon, which makes it easier to see.

A few examples:

{F1137554541} {F1137554533} {F1137554537} {F1137554549}

Reviewed By: muirdm

Differential Revision: D50897399

fbshipit-source-id: f7d25d404da009bdcd931d607c163d56b24c52b4