Add blue dot and tooltip to resubmit stack action when some commits out of sync
Summary:
If there are commits in a stack that have changed locally and need syncing, we can make the "resubmit stack" button indicate this as a way of upselling using that button.

We already have a "sync" button next to each commit that needs syncing, but the idea is that this resubmit stack button would affect all these commits at once, and is more noticeable than the sync button thanks to the blue dot. Plus, submitting the entire stack often fixes weird dependency changes on phabricator.

Hovering on the resubmit stack button shows a matching blue dot explaining that one diff needs syncing.

I could imagine us hiding the individual "sync" buttons when we are already showing the blue dot on the stack resubmit. For now, let's try out just having both. This is the most flexible, at the cost of taking up a bit more space.

Reviewed By: muirdm

Differential Revision: D50901710

fbshipit-source-id: 1c071e62f717068172fcea8b09f270cb1e1c859b