Add option to compare diff with last submit
Summary:
If a commit is out of sync, add a context menu item to load a comparison of the commit with the version that was last submitted for review.

This depends on the `--since-last-submit`  arg for `sl diff`. This seems to work relatively well in my testing. It works with `--rev` which is the most important thing. It seems to only include files changed within the diff and not general changes on master if you've rebased a commit.

If we find no difference in the comparison, then it may be that the commit was rebased (or restacked, if a parent commit changed). So we can customize our empty state to mention this.

There's no use offering this comparison for commits that we believe to be in sync. So we only show it for out of sync commits (out of sync in either direction).

Reviewed By: muirdm

Differential Revision: D50903086

fbshipit-source-id: 874785862c0c8c4bb70ee06135a1723465bdb2b9