Hide sync buttons if there's inline progress
Summary:
While submitting a commit (either singly or as part of a stack), we would still show the "sync" button. This is weird.

Instead, let's just hide the sync buttons while you have inline progress. This automatically means you won't be able to see the sync button during a submit.

Reviewed By: muirdm

Differential Revision: D50901891

fbshipit-source-id: 7d0d79765c64a587fcde586d2b2c4abf9ee8f4c2