chg: disable blackbox pre-fork
Summary:
We need to minimize the amount of stuff we do in Rust land before forking in chg. We have a lot of global state in Rust, and very little of it is designed to work properly after forking.

I noticed this code path was initializing the Rust thread rng pre-fork, but did not notice any other incorrect behavior. If something logged to blackbox pre-fork (without calling blackbox::init()), it would probably be duplicated in-memory in the children.

I will follow up with a specific fix for the thread local rng.

In general I don't think we should have to worry about forking. The longer term solution is to teach the Rust command server how to fork on unix, and use that instead of chg. The command server has a much smaller surface area for accidentally initializing something unsafe pre-fork.

Reviewed By: zzl0

Differential Revision: D51031697

fbshipit-source-id: 7ef1474980ecb341068ba028fa13bc9da393951a