Add request load logging on ODS
Summary:
Once a certain threshold of request load is reached we mark the task as unhealthy in Proxygen.
This should be a stopper metric for KFS as after that point increasing the weight of the task doesn't actuallly result in more traffic being sent to it.
I could add it to our KFS setup from scuba but it is less clean, and I think it's a metric that just makes less sense in the context of a request and more sense in the context of a task, so later I will also add this as an ods counter in Lfs and then remove it from Scuba.
In the future we should consider a setup for kfs that removes this healtchecks before overload testing.

Differential Revision: D51111143

fbshipit-source-id: 985989dc7264556af2f18287b5ef62c023d6068b