Support more granular FileChangeFilters in commit_transformation
Summary:
Extending the file changes filter type by supporting a way to specify **exactly when the filters are applied**.

This will be used to strip git submodules changes in the next diffs (D50967891).

The plan is: when git submodules should be stripped in a sync, the following filters will be applied:
1. Before implicit deletes only: only changes to git sub modules.
2. Before multi mover only (i.e. what should be synced): any change that is **NOT** to a git submodule.

The important filter is the 2nd, which will assure that any changes to git submodules are not synced.
However, the 1st filter should significantly improve performance, because it will only get implicit deletes caused by changes made to git submodules. We can do that because we know that in the scenarios we strip git submodules, the only implicit deletes that have a chance of not being synced, thus need to be expressed explicitly, are the ones caused by git submodules.

Reviewed By: mitrandir77

Differential Revision: D51064477

fbshipit-source-id: 2214f422b1a4c8feebee9d7c6afc3b158f200ab5