oss: make isl work on oss builds
Summary:
On our non-Windows OSS builds we are not including `isl-dist.tar.xz`, which makes `sl isl` not work.

This diff adds it directly on our `Makefile` rather than in `setup.py` as we already have the precedent of doing so for `make hg` and `setup.py` is not something we want to maintain in the long term.

Reviewed By: zzl0

Differential Revision: D51142970

fbshipit-source-id: 360240fc3c062b34726e48ededccdc31a78d4311