isl: allow more flexibilty when finding isl.tar
Summary:
Sometimes `sl` can fail to find `isl.tar.xz`. One reason for this is that it uses `sys.executable` for finding the "location" of the current executable. The problem is that this "location" can actually be just a symlink to the actual location; for instance, in Homebrew installations this can be a symlink:

```
$ ls -al $(which sl)
lrwxr-xr-x  1 sggutier  staff  53 Nov  8 14:35 /Users/sggutier/homebrew/bin/sl -> ../Cellar/sapling/0.2.20231107-153923+835a2a82/bin/sl
```

This makes changes so that it uses the actual location of the binary instead, as it is expected for `isl.tar.xz` to be in the same directory as `sl`.

Reviewed By: muirdm, zzl0

Differential Revision: D51134639

fbshipit-source-id: 42431642270bedd3bfafb03bc6ab3dba8a9cf3db