Refactor file change filter unit tests
Summary:
Small refactoring of the existing unit test for the file change filter, to reduce no-op changes made in D51064477 and make it easier to review.

I basically extract most of the unit test into an implementation function that I'll use to add more test cases in the next diff without copy-pasting a lot of code.

Differential Revision: D51110603

fbshipit-source-id: 0bdb4d7d94342da6663db4b8332aa3b2f1e3c7ca