fix HgDatapackStore use after free
Summary:
The most common crash is in the in the HgDatapackStore.

The crash traces to this access into the list of filtered paths
https://www.internalfb.com/code/fbsource/[0f21008428d409be301e60cef70d3eb82c02cc9f][history]/fbcode/eden/fs/store/hg/HgDatapackStore.cpp?lines=99

We use a reference into a value in the eden config after fetching finishes.
It's not safe to keep a reference to an object into an EdenConfig because
the config might be reloaded and the memory for that value freed. Even though
the code is blocking and thus it's safe to access references of local variables,
the eden config is not safe as another thread may invalidate the memory.

There are two potential solutions here: making the value it self ref counted
(wrapping the value in a shared pointer) or holding on to a reference to the
EdenConfig object. I keep going back-and-forth on which solution is better. I
think I slightly favor reference counting the value itself. I think its clearer
to need to keep a direct reference around than keep an indirect reference
around. This does mean an extra allocation when creating a new eden config, but
we already have all sorts of optimizations to avoid reloading the config often
and were gonna be allocating for this value anyways since it's a unordered_set.

Reviewed By: xavierd

Differential Revision: D51133312

fbshipit-source-id: 5d5a9592263474552db0efeb9e64e121481ac57f