prefetch: optimize the case with no base revision
Summary:
If you don't specifiy a "--base" revision, we diff the requested revision with the "null" revision which amounts to just walking all the files in the requested revision. The proper "walk" operation is a simpler and more efficient than "diff" (more parallelized), so let's use it in this case.

In particular, one issue with the "diff" implementation is it uses a 1ms batch window. If you are starved for CPU, you end up with batches of size 1 which does not work out very well. The "walk" implementation does not rely on a timeout to fill batches, so it won't suffer from this problem.

I added a new "walkfiles" method to pymanifest that yields the paths and the file nodes.

We can probably rework "diff" to improve parallelism, but that is a bit more work.

Reviewed By: xavierd

Differential Revision: D51170845

fbshipit-source-id: fc83c1cf9b0abd3f47c441b474d621fb35c85673