edenfs_config_manager: add scuba logging rollout names
Summary:
This diff adds scuba logging to `edenfs_rollouts` (in `infrastructure` namespace). This table is intentionally bare bones, since it can/should be `join`ed with `dev_command_timers` to look at trends. I plan to add a diaquery unidash in the future to make this easier - T168147059.

I explicitly only log `host` (and not `user`) since the script is usually ran as `root`.

Reviewed By: xavierd

Differential Revision: D50665152

fbshipit-source-id: 3d8de9a67c91d91b1481362be52070507acf2693