sharded_map_v2: implement from_entries and from_entries_and_partial_maps
Summary:
from_entries constructs a ShardedMapV2Node out of an iterator over key-value pairs.

from_entries_and_children constructs a ShardedMapV2Node from a TrieMap of values and partial maps. The key for every input partial map is a prefix that's prepended to it, which represents that keys that have this prefix are all contained in this child node. This method returns an error if the key for a partial map is a prefix of any other key in the input TrieMap.

Reviewed By: gustavoavena

Differential Revision: D50796566

fbshipit-source-id: e351e4eedfc277c59eac49f8072b292844c882d1