sharded_map_v2: convert i32s in unit tests to u32s
Summary: This is specifically a bug fix to the quickcheck tests as the default value used for MaxTestValue is 0, which means that we get an incorrect rollup max value if the generated test case contains only negative values, making the quickcheck test flaky. Alternatively I could fix this by making the default value be i32::MIN, but not having negative values makes the generated cases slightly simpler so I went with that instead.

Differential Revision: D51242911

fbshipit-source-id: e9a341406c12337fb56e2e09e8a57b16adeeede3