cross-repo sync: test cross-repo sync with different names (fail successfully)
Summary:
Write a test for cross-repo sync that exposes the current behaviour with respect to rewriting a bookmark that has a different name between the large and the small repo.

The current behaviour is incorrect as the `common_pushrebase_bookmarks_map` is ignored and the bookmark: `heads/main` in the small repo is treated exactly the same as any other bookmark instead of being considered to match `main` in the large repo.

Reviewed By: mitrandir77

Differential Revision: D51069866

fbshipit-source-id: 615519e3dfe7899a671a9436ad7ab787cf4a9c6c