workingcopy: attempt to avoid slow status during checkout
Summary:
When waiting for "working copy" changes, try to ensure that the wlock is not
being held by others. This might avoid slow status when a checkout is
in-progress.

Note: it does not completely avoid such slow status since the checkout can
still happen after the wlock check. A better solution might be making `status`
error out like edenfs during checkout. But that's a larger change.

Reviewed By: zzl0

Differential Revision: D50808547

fbshipit-source-id: d3159059915642dae5c34849bd4d295dd00196b9