status: add fallback to slow crawl on watchman failure
Summary:
The Python "status" code would fall back to non-Watchman status on a Watchman error by default. Previously I didn't reproduce this in the Rust status, hoping it wasn't necessary. However, various cases have popped up where Watchman can't be used in various environments for various reasons, so let's add the fallback to Rust as well.

Note that some of the Watchman errors have been real bugs in Watchman (e.g. startup race condition in Windows), and potential "environment" bugs (e.g. Watchman simply not installed), but it does seem like we should err on the side of not blocking users with errors. I'm going to disable fsmonitor.fallback-on-watchman-exception for tests and local development so we don't hide new issues.

Reviewed By: quark-zju

Differential Revision: D51265962

fbshipit-source-id: d66dd47e83f4bd3811f04d237a75148030cc8526