Accept non-Meta flavour MySQL in the metadata db config
Summary:
In this diff I make it possible to pass a non-Meta flavour MySQL databse to be used as a metadata database by Monononoke.
In order to do this I create a separate config in `eden/mononoke/metaconfig/types/src/lib.rs` which contains the info necessary to locate and connect to the database.
This config is consumed in `eden/mononoke/common/rust/sql_ext/src/facebook/impl/oss_mysql.rs`, which creates the necessary connections. Because the request to secret manager is async,  the async had to be propagated, touching other parts of the codebase.

Reviewed By: mitrandir77

Differential Revision: D51067227

fbshipit-source-id: ec9c88fd432c95370c6f80ae44376225a75519c2