remove prefetch logic from `prefetch-profile activate` command
Summary:
The current UX for the activate command is a bit confusing. The first time a user runs `prefetch-profile activate {profile}`, the command will run an initial prefetch for the given profile. This initial prefetch is somewhat unexpected (since the command is named `activate`, not `activate-and-prefetch`. To make things more confusing, there are options to skip the initial prefetch, but this isn't the default.

To make things more user friendly, we're going to make the activate command simply add the profile to a list of prefetch profiles that can be fetched on-demand. The user will then have to run `prefetch-profile fetch {profile}` to fetch files for that single profile, or they can run `prefetch-profile fetch` with no arguments to fetch all activated prefetch profiles.

The second case (fetch with no arguments) is currently broken. That will be fixed in the next diff.

Differential Revision: D51178613

fbshipit-source-id: 261ce368d9132f7694c88bbc34e93a3a15a50529