introduce new `prefetch-profile fetch` behavior
Summary:
The old `prefetch-profile fetch` command didn't quite work as expected. It would fetch all active profiles if additional profile arguments were omitted, but it was hard to tell this was happening. This is due to the fact that it would just kick off the prefetch in the background without any indication to the user that the command was successful and fetches were started.

The command was supposed to print the list of files that were accessed in the fetch, but that behavior was broken. I would also argue that printing out all the files that were fetched is not needed.

Therefore, this diff introduces the following changes:

* fetch doesn't print out the list of fetched files
* fetch prints out a human-readable list of profiles that fetches were kicked off for
* fetch offers a `--json` option to output a machine readable list of prefetch profiles that were fetched
* all the help texts are modified to match the current behavior of the command

Differential Revision: D51178612

fbshipit-source-id: 58f1f86f9def14ccc833b7e9e7e23e5d1f9a1281