add example to `eden prefetch-profile fetch` documentation
Summary:
The help text is a bit confusing. In some places, it was documented that -- was a required argument. The reality is that `--` is only required when a user provides both the --commits option and a list of profile arguments.

To make this more clear, add some extra info to the help text and add some examples of each use case.

Differential Revision: D51224940

fbshipit-source-id: 6077929f5228c1738458315f77115027e27c8aad