commit_graph: move the in-memory commit graph tests to testlib
Summary:
The in-memory commit graph is used by the commit graph testlib for some of the generated tests.  This creates a circuar dependency, as the in-memory commit graph tests also depend on the testlib.

Resolve this by moving the in-memory commit graph tests to the testlib.  Arguable they test both the in-memory commit graph and the testlib itself at the same time, so this seems fine.

Differential Revision: D51154998

fbshipit-source-id: c0f1a3b939f512fb1bee86b91d66c4199de24b80