commit_graph: add stats for memcache prefetch hits
Summary:
We are going to start storing prefetches in memcache, so that it's possible to hit against a prefetch peformed by another host in the region.

In this step, we add statistics for:

* When a memcache hit allows us to prefetch additional values (`mononoke.cache.commit_graph.prefetch.memcache_prefetched`).
* When a cachelib hit can be attributed to a memcache prefetch (`mononoke.cache.commit_graph.prefetch.memcache_hit`).

Note that in `fill_cachelib` we use the `insert_handle` method in cachelib, not the `insert_or_replace_handle`.  This is important, as it means the `CacheOrigin` field in the cachelib entry will be the *first* origin for a particular entry.  This means we will only count hits that can reasonably be attributed to the memcache hit, and not from any other prefetch operation.

Reviewed By: mitrandir77

Differential Revision: D51200007

fbshipit-source-id: 3162f0dd57a64ddbd6d5f168a4d97ac55eb98c8b