commit_graph: add prefetched edges to memcache commit graph edges
Summary:
We are going to start storing prefetches in memcache, so that it's possible to hit against a prefetch peformed by another host in the region.

In this step, we replace the values that are stored in the cache so that they can include prefetched edges.

For the cachelib cache, we do *not* want to store prefetched values, as they will just take up space and not be useful.  If a prefetch occurred, all of its entries will be in their own cachelib cache entry.

To make this possible, we need to manually implement the `Abomonation` trait in order to drop the prefetched values as they entombed and exhumed.

(IMPORTANT)**SAFETY:** Abomonation's processing is somewhat questionable, however we use the code that would be generated by the `Abomonation` proc macro as the basis for the implementations.  We make sure that the methods match what they do.

For the memcache cache, we change the data that is stored to be a new struct that includes the prefetched values.

NOTE: This is a format change for memcache, so we bump the memcache codever.

Reviewed By: mitrandir77

Differential Revision: D51154995

fbshipit-source-id: eb03df8a287fed93b48bb433a046c8e3fac3a8bb