commit_graph: return the origin commit for prefetched edges
Summary:
We are going to start storing prefetches in memcache, so that it's possible to hit against a prefetch peformed by another host in the region.

In this step, we add `Prefetched` as an alternative to `Fetched` for how changeset edges are fetched.   This also includes the original changeset that triggered the prefetch, so that we can associate the prefetched edge with it.

In the SQL storage, we track the original changeset id for each prefetched value, and use this to populate the prefetched values.

Reviewed By: mitrandir77

Differential Revision: D51154997

fbshipit-source-id: 59404f0796bfc21b15fcd79a2ce1498edd45cb8b