commit_graph: make tests flush the cache after building the graph
Summary:
Currently the tests all build the commit graph and then run the tests on it immediately.  In the case of the caching storage, this means the cache is pre-filled with all of the edges.  As a result, we don't actually test the cache filling process as part of the test.

Modify the tests so that we are able to flush the cache after the test has been set up but before we perform the testing.  This should exercise both the cache filling and hitting parts of the cache.

Differential Revision: D51154994

fbshipit-source-id: 6f8930a1a6f58c89f40ad520e535e61c3378583e