commit_graph: introduce FetchedChangesetEdges to track fetch origin
Summary:
We are going to start storing prefetches in memcache, so that it's possible to hit against a prefetch peformed by another host in the region.

In this step, we introduced `FetchedChangesetEdges` that is a wrapper around `ChangesetEdges` that can carry additional information about how it was fetched.  For now it has a single variant, as all edges are just fetched.

Differential Revision: D51154996

fbshipit-source-id: c505ee201c43878b0446c29d783e9f75b3a96d92