commit_graph: split memcache keygens based on prefetch type
Summary:
We are going to start storing prefetches in memcache, so that it's possible to hit against a prefetch peformed by another host in the region.

In this step, we will split the memcache keygens to partition the cache based on the type of prefetch.

We have two types of prefetch: linear p1, or skew-binary skip-tree.  We also need to account for no prefetch.  For each of these, generate a separate keygen, and use the appropriate keygen based on the request.

Reviewed By: mitrandir77

Differential Revision: D51154999

fbshipit-source-id: 89d86f6f56b4a6db25eb9cfdc59f1c3a7c20a238