Support GitSubmodulesChangesAction in commit_rewriting crate
Summary:
## This stack
This stack will update the forward syncer binary to support a new command (for now, called `initial-import`), which will import all commits from a small repo into a large repo and will **strip any changes to git submodules** based on the value of the `git_submodules_action` field in the small repo configs.

## This diff
- Creates the new `initial-import` command
- Starts using the `git_submodules_action` field from the small repo configs to filter out changes to any git submmodules.
  - It does so by using the `FileChangeFilters` that were extended in D51064477.

My implementation basically follows the same pattern as the multi mover generation. My line of though was: the logic for the multi mover is obtained from the sync config and the mover is passed to the `rewrite_commit` function. I had to find the right place to create the `FileChangesFilter`, so I went with the same approach.

While working on this diff I found that commit sync data provided isn't needed anymore. I created T169306120 to possibly delete it later though, because it seems to be a non-trivial refactoring that I didn't want to block this work stream.

## Important TODO
One important item that I'll have to cover before usable is what to do with commits that only change git submodules. I created T169315758 for this and will discuss this with a few people offline before implementing it.

Reviewed By: mitrandir77

Differential Revision: D50967891

fbshipit-source-id: ff7e9eda54ec6ff8c4ed9f873d8461f4ee20477d