Add config for amend restacking
Summary:
Add a UI setting corresponding to `amend.autorestack`. This controls how amending the middle of a stack behaves. By default, we will always restack, since that's what people using ISL are used to. Also, in the UI, it's a bit more clear about what conflicts you're in. On the CLI, it makes sense to avoid conflicts that aren't your direct amend, but here we might as well save you the trouble / confusion.

I updated this diff from a previous version. In this version:
- there's an ISL only localStorage config that controls this behavior
- it defaults to always autorestack
- it passes the config via the `--config` to the operation itself, as opposed to setting the amend.autorestack config in sl.

Reviewed By: quark-zju

Differential Revision: D51331264

fbshipit-source-id: 439f9eef936f21b013c0bf8ca783fa7c1046e48f