Specify multiple modifiers with array instead of bitwiseor
Summary: Previously, we specified modifiers with the bitwise or. This is convenient when testing for a match. But now we want to know the modifiers without extracting them from the union. So making them an array is convenient. We can still combine them with bitwise or when testing for equality.

Reviewed By: quark-zju

Differential Revision: D51326895

fbshipit-source-id: 73e79304ca47d72953f0fc824474aebf9d4e9279