Allow passing an event emitter to open a 'click' tooltip externally
Summary:
We'd like to provide keyboard shortcuts to open menu items. These current work as tooltips with `trigger='click'`. That doens't work with shortcuts well.

- we could try to listen for shortcuts inside <Tooltip>, but that seems like a weird place to put that control
- we could add a second tooltip that is keyboard controlled, but that leads to weird alignment issues and can cause tooltip duplication if you press the shortcut when it's already open

instead, I pass an event emitter to the tooltip, as a way of providing additional toggle triggers. This lets us convert keyboard events into the equivalent click to open the tooltip.

Reviewed By: quark-zju

Differential Revision: D51326892

fbshipit-source-id: d49d842a0a914e134e19b7b105681ea2a5e80298